Разбиение на модули
