+++ index.html
@@ -6,6 +6,7 @@
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
+  <script src="./js/main.js" type="module" defer></script>
   <title>Кекстаграм</title>
 </head>
 
@@ -234,6 +235,5 @@
       </section>
     </template>
 
-    <script src="./js/main.js" defer></script>
 </body>
 </html>
